// src/types/rada_gmin_types.ts
export interface GminaItem {
	"TERYT gminy": string;
	Gmina: string;
	Powiat: string;
	Województwo: string;
	Numer: string;
	Mieszkańcy: string;
	Wyborcy: string;
	"Wysłane pakiety wyborcze": string;
	Siedziba: string;
	Miejscowość: string;
	Ulica: string;
	"Numer posesji": string;
	"Numer lokalu": string;
	"Kod pocztowy": string;
	Poczta: string;
	"Typ obwodu": string;
	"Przystosowany dla niepełnosprawnych": string;
	"Typ obszaru": string;
	"Pełna siedziba": string;
	"Opis granic": string;
	"Liczba wyborców w części B spisu wyborów": string;
}